+++ CHANGELOG.md
@@ -25,6 +25,7 @@ paths. Default names are changed for PID-files, control sockets and log files.
 Default log file name for an instance is `tarantool.log`. `tarantool`'s
 stdout/stderr and `tt` logs go to `tt.log` file.
 - Remove URI with creds from console title and prompt.
+- Ignore app-instance delimiters for Tarantool 3.0 instances.
 
 ### Added